Allow database cleaner to run against ‘remote’ URLs in CI
The database host is not localhost, it is ‘mysql’ which is the name of the Docker service that is running the database.
Database::Cleaner refuses to run against things that don’t say “localhost” to protect you from
Deleting real data by accident. Overriding this to allow it to use the docker container for mysql/tests for CI.